Terminando de fazer a maior parte do crud

+++ src/main/java/com/controller/ServletDeletarUsuario.java
@@ -1,21 +1,27 @@
 package com.controller;
 
 import java.io.*;
-import java.util.List;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.*;
 import jakarta.servlet.annotation.*;
 import com.repository.FuncionarioDao;
-import com.model.Funcionario;
 
 @WebServlet(name = "servletDeletarUsuario", value = "/servletDeletarUsuario")
 public class ServletDeletarUsuario extends HttpServlet {
     public void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         int id = Integer.parseInt(request.getParameter("id"));
+        String email = request.getParameter("email");
+        String emailFuncionario = (String) request.getSession().getAttribute("emailfuncionario");
         FuncionarioDao funcionarioDao = new FuncionarioDao();
-        funcionarioDao.deleteById(id);
+        if(!email.equals(emailFuncionario)) {
+            funcionarioDao.deleteById(id);
+        }
+        System.out.println(request.getParameter("empresa"));
+        if(((String)request.getParameter("empresa")).equals("sim")) {
+            request.getRequestDispatcher("PaginaAposLogin/empresa.jsp").forward(request, response);
+        }
         request.getRequestDispatcher("PaginaAposLogin/crud.jsp").forward(request, response);
     }
 }
